cast freq to double for arm compatibility

+++ modules/test/benchmark/include/nt2/sdk/bench/metric/gflops.hpp
@@ -37,7 +37,7 @@ namespace nt2 { namespace bench
                       ) const
     {
       unsigned long freq = get_cpu_freq();
-      double m = ( e.flops() * e.size() * freq / (Stat::evaluate(c) * 1000000000 ));
+      double m = ( e.flops() * e.size() * (double)freq / (Stat::evaluate(c) * 1000000000. ));
       details::measures_map[details::identify_result(name,e,*this)] = m;
       return m;
     }
